# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Voiceover Tool API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ApiKey {
  id: number;
  label: string;
  /** Masked key e.g. "sk-...abc" */
  keyPreview: string;
  isActive: boolean;
  usageCount: number;
  /** @nullable */
  lastUsedAt?: string | null;
  createdAt: string;
}

export interface ApiKeyInput {
  /** @minLength 1 */
  label: string;
  /** @minLength 1 */
  key: string;
}

export interface ApiKeyUpdate {
  label?: string;
  isActive?: boolean;
}

export interface AdminStats {
  totalKeys: number;
  activeKeys: number;
  totalGenerations: number;
  totalCharacters: number;
}

export interface Voice {
  voiceId: string;
  name: string;
  category: string;
  /** @nullable */
  description?: string | null;
  /** @nullable */
  previewUrl?: string | null;
}

export interface TtsInput {
  /**
     * @minLength 1
     * @maxLength 5000
     */
  text: string;
  voiceId: string;
  /**
     * @minimum 0
     * @maximum 1
     */
  stability?: number;
  /**
     * @minimum 0
     * @maximum 1
     */
  similarityBoost?: number;
  modelId?: string;
}

export interface TtsResult {
  generationId: number;
  audioUrl: string;
  characterCount: number;
}

export interface Generation {
  id: number;
  text: string;
  voiceId: string;
  voiceName: string;
  characterCount: number;
  audioUrl: string;
  /** @nullable */
  modelId?: string | null;
  createdAt: string;
}

export interface GenerationList {
  items: Generation[];
  total: number;
}

export interface ErrorResponse {
  error: string;
}

export interface RegisterInput {
  /** @minLength 1 */
  name: string;
  email: string;
  /** @minLength 6 */
  password: string;
}

export interface LoginInput {
  email: string;
  password: string;
}

export interface AuthUser {
  id: number;
  name: string;
  email: string;
  createdAt: string;
}

export type ListGenerationsParams = {
limit?: number;
offset?: number;
};