# lib/api-zod/src/generated/types/authUser.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Voiceover Tool API
 * OpenAPI spec version: 0.1.0
 */

export interface AuthUser {
  id: number;
  name: string;
  email: string;
  createdAt: string;
}