# lib/api-zod/src/generated/types/registerInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Voiceover Tool API
 * OpenAPI spec version: 0.1.0
 */

export interface RegisterInput {
  /** @minLength 1 */
  name: string;
  email: string;
  /** @minLength 6 */
  password: string;
}